+++ ml/src/main/scala_spark_2.2/org/apache/spark/ml/h2o/SparkSpecificUtils.scala
@@ -19,37 +19,8 @@ package org.apache.spark.ml.h2o
 
 import org.apache.spark.SparkContext
 import org.apache.spark.ml.h2o.models.CrossSparkUtils
-import org.apache.spark.ml.param.Params
-import org.apache.spark.ml.util.DefaultParamsReader
-import org.apache.spark.ml.util.DefaultParamsReader.Metadata
-import org.json4s.jackson.JsonMethods._
-import org.json4s.{DefaultFormats, JObject}
 
 object SparkSpecificUtils extends CrossSparkUtils {
-
-  private def getAndSetParams(
-                               instance: Params,
-                               metadata: Metadata,
-                               skipParams: Option[List[String]] = None): Unit = {
-    implicit val format = DefaultFormats
-    metadata.params match {
-      case JObject(pairs) =>
-        pairs.foreach { case (paramName, jsonValue) =>
-          if (skipParams == None || !skipParams.get.contains(paramName)) {
-            val param = instance.getParam(paramName)
-            val value = param.jsonDecode(compact(render(jsonValue)))
-            instance.set(param, value)
-          }
-        }
-      case _ =>
-        throw new IllegalArgumentException(
-          s"Cannot recognize JSON metadata: ${metadata.metadataJson}.")
-    }
-  }
-
-  override def setTransformerParams(metadata: DefaultParamsReader.Metadata, instance: Params, filteredParams: List[String]): Unit = {
-    getAndSetParams(instance, metadata, Some(filteredParams))
-  }
 
   override def getNumTaskForStage(sc: SparkContext, stageId: Int): Int = {
     sc.jobProgressListener.stageIdToInfo(stageId).numTasks
